requirements.txt
alembic==1.13.2
aniso8601==7.0.0
annotated-types==0.7.0
anyio==4.4.0
asyncpg>=0.24,<1.0
certifi==2024.7.4
click==8.1.7
dnspython==2.6.1
email_validator==2.2.0
fastapi>=0.115.3
fastapi-cli==0.0.4
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
Mako==1.3.5
moviepy==1.0.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
phonenumbers==8.13.48
cloudinary==1.41.0
pre-commit==3.8.0
promise==2.3
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.9.0
pytest==8.3.2 
python-dotenv==1.0.1
requests==2.32.3
starlette==0.40.0
SQLAlchemy>=2.0.14
sqlmodel==0.0.21
strawberry-graphql==0.243.0
uvicorn==0.32.0